rest-server/requirements.txt
flask
flask-cors
requests
redis
google-cloud-pubsub
google-cloud-storage
